apps/telephony_app/pyproject.toml
[tool.poetry]
name = "telephony-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pyngrok = "^7.1.6"
python-dotenv = "^1.0.1"
vocode = { path = "../..", extras = [
    "all",
], develop = true, python = ">=3.11,<3.12" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
